Refactor script to use multiple classes
